+++ CBV-with-CBV-embedding/js/nodes/unop.js
@@ -22,23 +22,24 @@ class UnOp extends Node {
 	}
 
 	rewrite(token, nextLink) {
-		if (nextLink.to == this.key) {
-			if (token.rewriteFlag == RewriteFlag.F_OP) {
-				token.rewriteFlag = RewriteFlag.EMPTY;
-				var wrapper = BoxWrapper.create().addToGroup(this.group);
-				var newConst = new Const(token.dataStack.last()).addToGroup(wrapper.box);
-				var newLink = new Link(wrapper.prin.key, newConst.key, "n", "s").addToGroup(wrapper);
-				nextLink.changeTo(wrapper.prin.key, "s");
-				this.graph.findNodeByKey(this.findLinksOutOf(null)[0].to).delete();
-				this.delete();
+		if (token.rewriteFlag == RewriteFlag.F_OP && nextLink.to == this.key) {
+			token.rewriteFlag = RewriteFlag.EMPTY;
+			var wrapper = BoxWrapper.create().addToGroup(this.group);
+			var newConst = new Const(token.dataStack.last()).addToGroup(wrapper.box);
+			var newLink = new Link(wrapper.prin.key, newConst.key, "n", "s").addToGroup(wrapper);
+			nextLink.changeTo(wrapper.prin.key, "s");
+			this.graph.findNodeByKey(this.findLinksOutOf(null)[0].to).delete();
+			this.delete();
 
-				token.rewriteFlag = RewriteFlag.F_PROMO;
-				token.rewrite = true;
-				return newLink;
-			}
+			token.rewriteFlag = RewriteFlag.F_PROMO;
+			token.rewrite = true;
+			return newLink;
 		}
-		token.rewrite = false;
-		return nextLink;
+		
+		else if (token.rewriteFlag == RewriteFlag.EMPTY) {
+			token.rewrite = false;
+			return nextLink;
+		}
 	}
 
 	unOpApply(type, v1) {
